requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
feedparser==6.0.11
fonttools==4.57.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.0
h11==0.14.0
holidays==0.70
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.32
matplotlib==3.10.1
numpy==2.2.5
openai==1.75.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prophet==1.1.6
PuLP==3.1.1
py==1.11.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stanio==0.5.1
tavily-python==0.5.4
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0